CCF-11947 parse vcpus as a float

+++ src/commands/actions.js
@@ -157,7 +157,7 @@ module.exports.InvokeActionCommand = class {
         if (options.memory)
             params.properties['memory'] =parseInt(options.memory);
         if (options.vcpus)
-            params.properties['vcpus'] =parseInt(options.vcpus);
+            params.properties['vcpus'] =parseFloat(options.vcpus);
 
         // Set the API Endpoint and Token if not specified
         if (!params.apiEndpoint) params.apiEndpoint = profile.url;
